Modificato l'algortimo di calcolo delle Durate riscaldando quello originale in modo da avere valori espressi in unità 2msec ed essere paragonabili alle condizioni attuali. Rivista la logica dell'Accensione del Nebulizzatore e della Pompa: prima si attiva SOLO in Nebulizzatore per la Durata calcolata dall'Algoritmo, poi sia il Nebulizzatore che la Pompa insieme per la Stessa Durata. Quando lo Stato del Tappo passa da Aperto a Chiuso: prima si attiva SOLO in Nebulizzatore per la Durata calcolata dall'Algoritmo, poi sia il Nebulizzatore che la Pompa insieme per la Stessa Durata.

+++ Humidifier_Program/timerMg.c
@@ -26,7 +26,7 @@ unsigned short TimeBase;
 
 timerstype TimStr[N_TIMERS];
 
-unsigned short Durata[N_TIMERS] = {
+unsigned long Durata[N_TIMERS] = {
    /* 1 */ DELAY_READ_IO,
    /* 2 */ DELAY_INTRA_FRAMES,
    /* 3 */ DELAY_WAIT_SLAVE,
